+++ src/main/java/himedia/photobook/controllers/admin/AdminController.java
@@ -1,9 +1,7 @@
 package himedia.photobook.controllers.admin;
 
 import org.springframework.stereotype.Controller;
-
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 @RequestMapping({"/admin"})
@@ -11,11 +9,6 @@ public class AdminController {
 	@RequestMapping({"","/home"})
 	public String home() {
 		return "/WEB-INF/views/admin/admin_pages.jsp";
-	}
-	
-	@RequestMapping({"/customerManagement", "/customerManage", "/cm"})
-	public String customerManagement() {
-		return "/WEB-INF/views/admin/admin_customer_management.jsp";
 	}
 	
 	@RequestMapping({"/orderManagement", "/orderManage", "/om"})
@@ -34,15 +27,5 @@ public class AdminController {
 	@RequestMapping({"/dashboard", "/dash"})
 	public String dashboard() {
 		return "/WEB-INF/views/admin/admin_dashboard.jsp";
-	}
-
-	@GetMapping("/update")
-	public String customerUpdate() {
-		return "/WEB-INF/views/admin/admin_customer_update.jsp";
-	}
-	
-	@GetMapping("/delete")
-	public String customerDelete() {
-		return "/WEB-INF/views/admin/admin_customer_delete.jsp";
 	}
 }
