+++ src/main/java/himedia/photobook/controllers/admin/AdminCustomerController.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -20,6 +19,11 @@ public class AdminCustomerController {
 	@Autowired
 	private AdminCustomerService adminCustomerService;
 	
+	@GetMapping({"/customerManagement", "/customerManage", "/cm"})
+	public String customerManagement() {
+		return "/WEB-INF/views/admin/admin_customer_management.jsp";
+	}
+	
 	@GetMapping("/search")
 	public String searchUSers(@RequestParam Map<String, String> params, Model model) {
 		String searchCategory = params.getOrDefault("search-category", null);
@@ -29,17 +33,24 @@ public class AdminCustomerController {
 		return "/WEB-INF/views/admin/admin_customer_management.jsp";
 	}
 	
-	@GetMapping("/confirmDelete")
-	public String confirmDelete(@RequestParam ("userId") String userId, Model model) {
-		UsersVo users = adminCustomerService.getUserById(userId);
-		model.addAttribute("users", users);
-		return "/WEB-INF/views/admin/admin_confirmDelete.jsp";
+	@GetMapping("/update")
+	public String customerUpdate() {
+		return "/WEB-INF/views/admin/admin_customer_update.jsp";
 	}
 	
-	@PostMapping("/delete")
+//	관리자 고객관리 고객목록 삭제확인페이지(구현X)	
+//	@GetMapping("/customerDelete")
+//	public String customerDelete(@RequestParam ("userId") String userId, Model model) {
+//		UsersVo users = adminCustomerService.getUserById(userId);
+//		model.addAttribute("users", users);
+//		return "/WEB-INF/views/admin/admin_customer_delete.jsp";
+//	}
+	
+	@GetMapping("/delete")
 	public String deleteUser(@RequestParam("userId")String userId) {
+		System.out.println("deleteUser userId: " + userId);
 		adminCustomerService.deleteUser(userId);
-		return "redirect:/admin/userList";
+		return "redirect:/admin/cm";
 	}
 
 }
