+++ src/main/resources/mybatis/mappers/users.xml
@@ -49,12 +49,12 @@
         	ADDRESS as address
         FROM USERS
         WHERE
-        	USER_ID = #{userId}
+        	USER_ID = #{value}
     </select>
     
-    <delet id="deleteUser" parameterType="string">
+    <delete id="deleteUser" parameterType="string">
     	DELETE FROM USERS
-    	WHERE userId = #{userId}
-    </delet>
+    	WHERE USER_ID = #{value}
+    </delete>
    
 </mapper>
